+++ asserto/_asserto.py
@@ -23,6 +23,7 @@ from .handlers import StringHandler
 
 # Todo: base: `tidy up docstrings`
 # Todo: base `remove duplication here`
+# Todo: Api feels cumbersome with decorators; can we improve DRY-ness?
 
 
 class Asserto(AssertoBase):
@@ -222,8 +223,18 @@ class Asserto(AssertoBase):
     @handled_by(BaseHandler)
     def is_true(self) -> Asserto:
         """
-        Checks the actual value is True.
-        :return: The `Asserto` instance for fluency.
+        Asserts that the value provided as the actual value (self.actual) is explicitly referencing
+        the single `True`.  This does NOT perform a 'truthy' check but instead the value is checked
+        using the `is` identity check.  In order to check if a value is considered 'truthy' use the
+        `.truthy()` method instead.
+        """
+        return self._dispatch()
+
+    @handled_by(BaseHandler)
+    def is_truthy(self) -> Asserto:
+        """
+        Asserts that the truthy-ness of the value provided as the actual value (self.actual) is `True`
+        when converted to a boolean value.
         """
         return self._dispatch()
 
@@ -234,6 +245,10 @@ class Asserto(AssertoBase):
         :return: The `Asserto` instance for fluency.
         """
         return self._dispatch()
+
+    @handled_by(BaseHandler)
+    def is_falsy(self) -> Asserto:
+        """ Todo: Docs."""
 
     @handled_by(BaseHandler)
     def is_equal_to(self, other: typing.Any) -> Asserto: